requirements.txt
openai-whisper
sounddevice
numpy
scipy
pyttsx3
python-dotenv
torch
requests
google-generativeai
